kind of a plan to refactor boot.js to be more generic
pseudo code and CAPS ftw
